% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_density.R
\name{nba_density}
\alias{nba_density}
\title{NBA Schedule Density}
\usage{
nba_density(df)
}
\arguments{
\item{df}{A dataframe. The object storing the result of `nba_travel()`}
}
\value{
A data frame with the following columns:
 \describe{
         \item{Season}{A chracter string. The season(s) downloaded.}
         \item{Phase}{A character String. The phase of the season. RS or PO.}
         \item{Date}{Date object. The date of the game.}
         \item{Team}{A character String. The name of the team.}
         \item{Opponent}{A character String. The name of the opponent.}
         \item{Location}{A character string. Location of the game. (Home or Away)}
         \item{`W/L`}{A character string. Outcome of the game. (W or L)}
         \item{`B2B`}{A character string. B2B-1st (first game in B2B games) or B2B-2nd (second game in B2B games.)}
         \item{`B2B-1st`}{A character string. If yes, it is the first game of a back-to-back.}
         \item{`B2B-2nd`}{A character string. If yes, it is the second game of a back-to-back.}
         \item{`3in4`}{A character string. If yes, it is the third game in 4 days.}
         \item{`4in5`}{A character string. If yes, it is the fourth game in 5 days.}
         \item{`5in7`}{A character string. If yes, it is the fith game in 7 days.}
    }
}
\description{
Accepts the result of `nba_travel()` & returns a dataframe with some common arbitrary schedule density related metrics such as back-to-back(b2b), 3in4, 4in5, etc.
}
\examples{
datos <- nba_travel()
nba_density(df = datos)

}
