% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_injuries.R
\name{nba_injuries}
\alias{nba_injuries}
\title{NBA Injury List}
\usage{
nba_injuries(
  start_date = "2017-01-01",
  end_date = "2018-01-01",
  player = "",
  team = ""
)
}
\arguments{
\item{start_date}{Date. Minimum date to query data from. (i.e. 2010-12-03). It defaults to 2017-01-01 if not specified.}

\item{end_date}{Date. Maximum date to query data to. (i.e. 2018-12-14). It defaults to 2018-01-01 if not specified.}

\item{player}{Character String. Filter data by a specific player. (i.e. Pau Gasol). Note that if only the last name is indicated (i.e. Gasol) it will return results for all players with the same last name. It defaults to blank if not indicated.}

\item{team}{Character String. The nickname of a unique team to query. (i.e. Lakers). It defaults to blank (all teams) if not indicated.}
}
\value{
A data frame with the following columns:
 \describe{
         \item{Date}{A date. The date of the transaction.}
         \item{Team}{A character String. The nickname of the team where the transaction occurred.}
         \item{Acquired}{A character String. The name of the player involved in the transaction. Usually return to lineup.}
         \item{Relinquished}{A character String. The name of the player involved in the transaction. Usually pulled out of the lineup.}
         \item{Notes}{A character String. A short description about the transaction.}

    }
}
\description{
Returns a dataframe with injury transactions. This function queries data from <https://www.prosportstransactions.com/>
}
\examples{
nba_injuries(start_date = "2012-01-01",
             end_date = "2014-01-01",
             player = "Pau Gasol",
            team = "Lakers")


}
