% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_player_travel.R
\name{nba_player_travel}
\alias{nba_player_travel}
\title{NBA Individual Player Schedule, Travel Metrics and Game Stats}
\usage{
nba_player_travel(
  start_season = 2018,
  end_season = 2020,
  team = NULL,
  player = NULL,
  return_home = 20,
  phase = c("RS", "PO"),
  flight_speed = 450
)
}
\arguments{
\item{start_season}{Numeric. The year of the season users wish to explore (i.e. 2018)}

\item{end_season}{Numeric. The year of the final season users wish to explore (i.e. 2020)}

\item{team}{Character String. The name of the team to be explored. If empty it defaults to all teams within the selected seasons.}

\item{player}{Character String. The name of the player to be explored. If empty it defaults to all players in the selected query.}

\item{return_home}{Numeric. Users can set the number of days after which the team will return home between consecutive road games. It defaults to 20 if not indicated.}

\item{phase}{Character String. The phase of the season users wish to download. RS for Regular Season and PO for Playoffs. It defauls to both if not indicated.}

\item{flight_speed}{Numeric. Average cruising speed (in mph) for commercial flights. Used to adjust flight time estimation.}
}
\value{
A data frame with the following columns:
 \describe{
         \item{Season}{A chracter string. The season(s) downloaded.}
         \item{Phase}{A character String. The phase of the season. RS (Reg Season) or PO (Play-offs).}
         \item{Date}{A date. The date of the game.}
         \item{Team}{A character String. The name of the team.}
         \item{Opponent}{A character String. The name of the opponent.}
         \item{Location}{A character string. Location of the game. (Home or Away)}
         \item{`W/L`}{A character string. Outcome of the game. (W or L)}
         \item{City}{A character String. The name of the city in which the game is played.}
         \item{Distance}{Numeric. Estimated distance travelled (in miles) prior to a game.}
         \item{Route}{A character String. The route travelled for each game. If no travel involved it defaults to "No Travel".}
         \item{Rest}{Numeric. The number of days between games. It defaults to 15 for the first game for each team. If the team is making a "return_home" trip it will default to NA.}
         \item{TZ}{A character string. The time zone where the game is played.}
         \item{`Shift (hrs)`}{Numeric. The time zoen change in hours.}
         \item{`Flight Time`}{A character string. Aproximate duration of the flight for a given trip.}
         \item{`Direction (E/W)`}{A character string. Travel direction for trips involving zone changes.}
         \item{`Return Home`}{A character string. If yes, team is making a "return_home" trip.}
         \item{Player}{A character string. The name of the player.}
         \item{`Team Rest``}{Numeric. Number of rest days for the team.}
         \item{`Player Rest`}{Numeric. Number of rest days for each player.}
         \item{`Games Played`}{Numeric. Number of games played by each player.}
         \item{MINs}{Numeric. Minutes played by each player.}
         \item{PTS}{Numeric. Points scored by each player.}
    }
}
\description{
Returns a dataframe with computed travel and schedule metrics for NBA teams and each player of the team. It requires the game_logs() function from on the nbastatR package written by Bresler, A (2020) <https://github.com/abresler/nbastatR> to query season schedule information that is needed to calculate travel metrics.
}
\examples{
nba_player_travel(start_season = 2018,
                  end_season = 2020,
                  return_home = 4,
                  team = "Cleveland Cavaliers",
                  player = "Jose Calderon")



}
