% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_travel_plot.R
\name{nba_travel_plot}
\alias{nba_travel_plot}
\title{NBB Flight Paths}
\usage{
nba_travel_plot(
  data,
  season = NULL,
  team = NULL,
  land_color = "#17202a",
  land_alpha = 0.6,
  city_color = "cyan4",
  city_size = 0.8,
  path_curvature = 0.05,
  path_color = "#e8175d",
  path_size = 0.5,
  title = "NBA Flight Paths",
  title_color = "white",
  title_size = 20,
  caption = "",
  caption_color = "gray",
  caption_size = 8,
  caption_face = "italic",
  major_grid_color = "transparent",
  minor_grid_color = "transparent",
  strip_text_size = 8,
  strip_text_color = "white",
  strip_fill = "transparent",
  plot_background_fill = "#343e48",
  panel_background_fill = "transparent",
  ncolumns = 6
)
}
\arguments{
\item{data}{A data frame. The result of nba_travel()}

\item{season}{Numeric. A four year digit (i.e. 2018). If not indicated it defaults to the most recent season in the dataset.}

\item{team}{A character string. The name of a team (or vector of teams) to display. If not indicated it will default to all teams.}

\item{land_color}{A character string. Color of the land.}

\item{land_alpha}{Numeric. The alpha value of the land between 0-1. Used to set transparecy level.}

\item{city_color}{A character String. The color of the points representing the cities.}

\item{city_size}{Numeric. The size of the point representing the cities.}

\item{path_curvature}{Numeric. The level of curvature for the path representing the flight. Defaults to 0.05.}

\item{path_color}{A character String. The color of the lines representing the flight path.}

\item{path_size}{Numeric. The thickness of the lines representing the flight paths. Defaults to 0.5.}

\item{title}{A character string. The plot title.}

\item{title_color}{A character string. The color of the title text.}

\item{title_size}{Numeric. The size of the title text. Defaults to 20.}

\item{caption}{A character string. The plot caption.}

\item{caption_color}{A character string. The color of the caption text.}

\item{caption_size}{Numeric. The size of the caption text. Defaults to 8.}

\item{caption_face}{A character string. TThe face of the caption text. Defaults to "italic".}

\item{major_grid_color}{A character string. The color of the major gridlines. Defaults to "transparent".}

\item{minor_grid_color}{A character string. The color of the minor gridlines. Defaults to "transparent".}

\item{strip_text_size}{Numeric. The size of the strip text. Default to 8.}

\item{strip_text_color}{A character string. The color of the strip text.}

\item{strip_fill}{A character string. The color of the strip. Defaults to "transparent.}

\item{plot_background_fill}{A character string. The color of the plot background.}

\item{panel_background_fill}{A character string. The color of the panel background.}

\item{ncolumns}{Numeric. The number of columns for facetted plots.}
}
\value{
A ggplot object
}
\description{
Plots the estimated flight paths for the selected season and team(s). It returns a ggplot object that can be further customized by the user.
}
\examples{
datos <- nba_travel(season = 2015:2018)
nba_travel_plot(data = datos,
                season = 2017,
                team = c("Chicago Bulls", "Miami Heat"),
                city_color = "white",
                plot_background_fill = "black",
                land_color = "gray",
                caption_color = "lightblue",
                ncolumns = 1)

}
